Modularized code. Split into subtabs and dedicated controllers
