+++ hawtio-web/src/main/webapp/app/runtime/html/layoutRuntime.html
@@ -1,0 +1,8 @@
+<ul class="nav nav-tabs connected" ng-controller="Runtime.NavController">
+  <li ng-repeat="link in breadcrumbs" ng-show="isValid(link)" ng-class='{active : isActive(link.href)}'>
+    <a ng-href="{{link.href}}{{hash}}" ng-bind-html-unsafe="link.content"></a>
+  </li>
+</ul>
+<div class="row-fluid">
+  <div ng-view></div>
+</div>
