--- hawtio-web/src/main/webapp/app/runtime/html/runtime.html
@@ -1,68 +1,0 @@
-<div class="runtime-controller"
-	ng-controller="Runtime.RuntimeController">
-
-	<div class="row-fluid">
-		<div class="span10">
-			<div class="expandable open">
-				<div title="Overview" class="title">
-					<i class="expandable-indicator"></i> Overview
-				</div>
-				<div class="expandable-body well">
-					Running: (pid: <b>{{pid}}</b>) {{runtime.VmName}} - {{runtime.SpecVersion}} ({{runtime.VmVersion}})
-					by {{runtime.VmVendor}} in <b>{{runtime.SystemProperties['user.name']}}@{{host}}</b> running ({{operatingSystem.Name}} version {{operatingSystem.Version}} with {{operatingSystem.AvailableProcessors}} * {{operatingSystem.Arch}} CPU cores)
-					since <b>{{runtime.StartTime | date: 'yyyy-MM-dd hh:mm:ss' }} (uptime: {{uptime}})</b>
-				</div>
-			</div>
-			<div class="expandable closed">
-				<div title="Command Line" class="title">
-					<i class="expandable-indicator"></i> Command line (simulated)
-				</div>
-				<div class="expandable-body well breakLong" zero-clipboard data-clipboard-text="{{javaPath}} -classpath {{classPath}} {{vmArgs}} {{runtime.SystemProperties['sun.java.command']}}" 
-				tooltip="minimal command line (add system properties from below if not specific enough)">
-					{{javaPath}} -classpath {{classPath}} {{vmArgs}} <strong>{{runtime.SystemProperties['sun.java.command']}}</strong>
-				</div>
-			</div>
-			<div class="expandable closed">
-				<div title="Working directory" class="title">
-					<i class="expandable-indicator"></i> Working directory
-				</div>
-				<div class="expandable-body well" zero-clipboard data-clipboard-text="{{workingDirectory}}">
-					<a href="file://{{workingDirectory}}">{{workingDirectory}}</a></b>
-				</div>
-			</div>
-			<div class="expandable closed" >
-				<div title="System Properties" class="title" >
-					<i class="expandable-indicator"></i> Key Metrics
-				</div>
-				<div class="expandable-body">
-					<div class="control-group inline-block pull-right">
-						<input type="text" class="search-query ng-pristine ng-valid"
-							placeholder="Filter..."
-							ng-model="metricsTableConfig.filterOptions.filterText"> <i
-							class="icon-remove clickable" title="Clear filter"
-							ng-click="metricsTableConfig.filterOptions.filterText = ''"></i>
-					</div>
-					<table class="table table-condensed table-striped"
-						hawtio-simple-table="metricsTableConfig"></table>
-				</div>
-			</div>
-			
-			<div class="expandable closed" >
-				<div title="System Properties" class="title" >
-					<i class="expandable-indicator"></i> System Properties
-				</div>
-				<div class="expandable-body">
-					<div class="control-group inline-block pull-right">
-						<input type="text" class="search-query ng-pristine ng-valid"
-							placeholder="Filter..."
-							ng-model="systemPropertiesTableConfig.filterOptions.filterText"> <i
-							class="icon-remove clickable" title="Clear filter"
-							ng-click="systemPropertiesTableConfig.filterOptions.filterText = ''"></i>
-					</div>
-					<table class="table table-condensed table-striped"
-						hawtio-simple-table="systemPropertiesTableConfig"></table>
-				</div>
-			</div>
-		</div>
-	</div>
-</div>
