+++ hawtio-web/src/main/webapp/app/runtime/html/systemProperties.html
@@ -1,0 +1,17 @@
+<div class="runtime-controller"
+	ng-controller="Runtime.SystemPropertiesController">
+
+	<div class="row-fluid">
+		<div class="span10">
+			<div class="control-group inline-block pull-right">
+				<input type="text" class="search-query ng-pristine ng-valid"
+					placeholder="Filter..."
+					ng-model="systemPropertiesTableConfig.filterOptions.filterText">
+				<i class="icon-remove clickable" title="Clear filter"
+					ng-click="systemPropertiesTableConfig.filterOptions.filterText = ''"></i>
+			</div>
+			<table class="table table-condensed table-striped"
+				hawtio-simple-table="systemPropertiesTableConfig"></table>
+		</div>
+	</div>
+</div>
