+++ hawtio-web/src/main/webapp/app/runtime/js/navbar.ts
@@ -1,0 +1,11 @@
+/**
+ * @module Diagnostics
+ */
+/// <reference path="./runtimePlugin.ts"/>
+module Runtime {
+
+  _module.controller("Runtime.NavController", ["$scope", "$location", "workspace", ($scope, $location, workspace) => {
+    Runtime.configureScope($scope, $location, workspace);
+  }]);
+
+}
